operation-server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
celery==5.3.6
flower==2.0.1
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
kafka-python==2.0.2
python-dateutil==2.8.2
pytz==2024.1
elasticsearch==8.11.0
